moved dao/model for product module
