+++ frontapi/src/main/java/com/harmoni/frontapi/main/product/dao/DevProductDaoService.java
@@ -1,6 +1,6 @@
-package com.harmoni.frontapi.main.product.service.dao;
+package com.harmoni.frontapi.main.product.dao;
 
-import com.harmoni.frontapi.main.product.service.model.Product;
+import com.harmoni.frontapi.main.product.model.Product;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
