+++ frontapi/src/main/java/com/harmoni/frontapi/main/product/model/Product.java
@@ -1,4 +1,4 @@
-package com.harmoni.frontapi.main.product.service.model;
+package com.harmoni.frontapi.main.product.model;
 
 import lombok.*;
